Courier-unicode and maildrop  formula
Added new formulas to complement imapsync tool. Due to some append error, imapsync documentation suggest to solve the issue running the reformime program, which is available on Linux but not on brew. The program is part of the maildrop packages, which requires courier-unicode library to properly work. Thus, the need of both.
The actual setup requires GCC too: for a reason I cannot understand, within brew environment clang fails. I tested it outside and looked good so maybe I still miss something